feat: Add embeddedInterpreterPath configuration option
This commit adds the embeddedInterpreterPath configuration option to the typhoon-test.testRun section in the package.json file. This option allows users to specify the path to the embedded Python interpreter. The updateEmbeddedInterpreterPath command has also been added to update the embeddedInterpreterPath configuration value. This change enhances the flexibility and customization of the Typhoon Test extension.