refactor: Update config.ts and pytestBuilder.ts
This commit updates the config.ts file to add a new function, refreshConfigs(), which refreshes the configuration values for the Typhoon Test extension. It also updates the pytestBuilder.ts file to modify the build() function, adding the -v flag to enable verbose output when running pytest. These changes improve the functionality and user experience of the Typhoon Test extension.