feat: Add Pytest Monitor view to TestTreeProvider
This commit adds the Pytest Monitor view to the TestTreeProvider in order to display real-time test monitoring in the tree view. By integrating the Pytest Monitor functionality, users can now monitor the progress and status of their tests as they run. This enhancement improves the usability and effectiveness of the Typhoon Test extension.