refactor: Update PytestArgumentBuilder to use pdfBuilder instance
This commit updates the PytestArgumentBuilder class in the src/models/PytestArgumentBuilder.ts file. It introduces a private instance of the PdfArgumentBuilder class called pdfBuilder. The getCommand() method in the getPDFConfig() function is updated to use the pdfBuilder instance instead of creating a new PdfArgumentBuilder object. Additionally, a new method getDisplayCommand() is added to PytestArgumentBuilder, which returns the display command by replacing the pdfConfig with the display command from the pdfBuilder instance.

Refactor the PytestArgumentBuilder to improve code readability and reuse the pdfBuilder instance.